MqttHandler/app/requirementsMqtt.txt
simplejson == 3.17.6
paho-mqtt == 1.6.1
